Implement UrlSet.pprint method
This patch adds a working pprint method to urlset objects. The point
being that you can now create an empty UrlSet container and fill it with
UrlSetelEments and then run the UrlSet.pprint() routine on a file object and
produce a valid sitemap file. I do not currently have an implementation
of pprint that works if we already have a handle.